backend/requirements.txt
fastapi==0.111.0 # Or the latest stable version
uvicorn[standard]==0.29.0 # Or the latest stable version
python-firebase-admin==6.3.0 # Or the latest stable version
pydantic==2.7.4 # Or the latest stable version (FastAPI dependency)